banners: AnythingButAnalogBanner
Port components/banners/anything-but-analog-banner.tsx → Svelte 5.
Uses VoronoiCanvas with the album art image; conditionally wraps in
an <a> when href prop is provided.
